requirements.txt
Django==2.1
djangorestframework==3.8.2
mysqlclient==1.3.13
pkg-resources==0.0.0
pytz==2018.5
